Overall tracking sheet

This function display the overall tracking sheet.

assessmentTrackingSheet(
  ds = NULL,
  sf = NULL,
  dssite = NULL,
  sfsite = NULL,
  survey_consent = NULL,
  sftarget = NULL,
  sfnbpts = NULL,
  formul = NULL,
  colorder = NULL
)

Arguments

ds
dataset as a data.frame object

sf
sampling frame as a data.frame object

dssite
name as a string of the field in the dataset where the site is stored

sfsite
name as a string of the field in the sampling frame where the site is stored

survey_consent
name as a string of the field in the dataset where the survey consent is stored

sftarget
name as a string of the field where the target number of survey is stored in the sampling frame

sfnbpts
name as a string of the field where the number of points generated is stored in the sampling frame

formul
formulas as a list of string used to compute the final number of eligible surveys and the variance from the target (C('formula1','formula2')). the values/fields available are: done and the ones generated according the survey consent values (one per value)

colorder
column names as a list of string to order the colums in the result (C('col1','col2',...)). the columns available are: site, done, final, variance and the ones generated according the survey consent values (one per value)

Value

ret_log the report

Examples

{
ds <- HighFrequencyChecks::sample_dataset
sf <- HighFrequencyChecks::SampleSize
dssite <- "union_name"
sfsite <- "Union"
survey_consent <- "survey_consent"
sftarget <- "SS"
sfnbpts <- "TotPts"
#formul <- c("done-no-not_eligible-deleted","done-no-not_eligible-deleted-SS")
#colorder <- c("site","SS","Provision","done","not_eligible","no","deleted",
#              "yes","final","variance")
formul=c("done-no-not_eligible","done-no-not_eligible-SS")
colorder=c("site","SS","TotPts","done","not_eligible","no","yes","final","variance")

list[dst,ret_log,var,graph] <- assessmentTrackingSheet(ds,
                        sf,
                        dssite,
                        sfsite,
                        survey_consent,
                        sftarget,
                        sfnbpts,
                        formul,
                        colorder)
head(ret_log,10)
}

#>                 site  SS TotPts done not_eligible no yes final variance
#> 1        baharchhara 253    317   19            7  0  12    12     -241
#> 2      haldia palong 260    325   27            5  1  21    21     -239
#> 3       jalia palong 259    324   19            5  1  13    13     -246
#> 4             nhilla 259    324   54            9  3  42    42     -217
#> 5       palong khali 255    319  115           36  7  72    72     -183
#> 6        raja palong 261    327   37            8  0  29    29     -232
#> 7       ratna palong 252    315   31           10  2  19    19     -233
#> 8            sabrang 260    325   20            8  1  11    11     -249
#> 9             teknaf 259    324   74           28  5  41    41     -218
#> 10 teknaf paurashava 253    317    4            1  0   3     3     -250