Check that the duration relative to the household size of each interview is more than a threshold

This function check that the duration relative to the household size of each interview is more than a specified threshold. There is an option to automatically mark for deletion the surveys which are under the threshold. Warning: If there are uncorrected mistakes in the survey dates, it can lead to have the length of the survey in seconds and this check will not performed well

chk5c_duration_Xmin_HHSize(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  HHSize = NULL,
  reportingcol = NULL,
  minduration = 10,
  delete = NULL
)

Arguments

ds
dataset as a data.frame object

survey_consent
name as a string of the field in the dataset where the survey consent is stored

dates
fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))

HHSize
name as a string of the field in the dataset where the household size is stored

reportingcol
columns as a list of string name from the dataset you want in the result (c('col1','col2',...))

minduration
minimum acceptable survey duration as integer in minutes

delete
delete action to be done as a boolean (TRUE/FALSE)

Value

ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors list of the errors found

Author

Yannick Pascaud

Examples

if (FALSE) {
df <- sample_dataset
sc <- "survey_consent"
hs <-"consent_received.respondent_info.hh_size"
dt <- c("survey_start","end_survey")
rc <- c("enumerator_id","X_uuid")
md <- 30
dl <- FALSE

chk5c_duration_Xmin_HHSize(df, sc, dt, hs, rc, md, dl)
}