Surveys that do not end on the same day as they started

This function check that all interviews in the dataset start and end the same day. There is an option to automatically mark for deletion the surveys which have different starting and ending dates.

isSurveyOnMoreThanADay(
  ds = NULL,
  survey_consent = NULL,
  dates = NULL,
  reportingcol = NULL,
  delete = NULL
)

Arguments

ds
dataset as a data.frame object

survey_consent
name as a string of the field in the dataset where the survey consent is stored

dates
fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))

reportingcol
columns as a list of string name from the dataset you want in the result (c('col1','col2',...))

delete
delete action to be done as a boolean (TRUE/FALSE)

Value

ds same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

errors list of the errors found

Examples

{
ds <- HighFrequencyChecks::sample_dataset
survey_consent <- "survey_consent"
dates <- c("survey_start","end_survey")
reportingcol <- c("enumerator_id","X_uuid")
delete <- FALSE

list_date_mistake <- isSurveyOnMoreThanADay(ds, survey_consent,dates, reportingcol, delete)
head(list_date_mistake[[2]], 10)
}

#>    enumerator_id                               X_uuid
#> 11            91 13d4e232-ba59-4be6-84b4-f5e7d103f549
#> 13            45 3f0d915b-7fa6-4262-8783-8d2c477af2e3
#> 14         30022 93ea674d-29c4-4750-bb48-ce2125a55094
#> 38         30019 8148d291-1918-43dd-b292-0932aea85f02
#> 45             3 2ab24867-db88-40f2-ba97-b5c7b6da8f1d
#> 46            76 0a3bcffb-0cba-4661-ac98-3c28b12b4ec8
#> 47         10021 5fb8b7e7-2692-47f3-8e83-98e4578ea3ce
#> 50            45 838d58ad-497b-4f4a-8392-b73c56a290c0
#> 54            93 7adab422-2c20-4552-beaf-16f1aa831b82
#> 57            65 7d6031c7-0dca-476b-89bc-2c5737f0e199
#>                     survey_start                    survey_end
#> 11    2018-11-17T10:54:27.159+06    2018-11-17T11:34:48.907+06
#> 13    2018-11-22T10:12:47.885+06    2018-11-22T11:06:02.342+06
#> 14    2018-11-17T10:32:06.261+06    2018-11-17T11:30:13.108+06
#> 38    2018-12-02T10:21:11.845+06    2018-12-02T11:19:50.998+06
#> 45    2018-11-29T10:07:44.092+06    2018-11-29T11:07:05.560+06
#> 46    2018-12-02T10:43:32.268+06    2018-12-02T11:08:19.101+06
#> 47    2018-12-01T10:59:07.195+06    2018-12-01T11:51:14.237+06
#> 50 2018-11-29T10:39:12.023+06:30 2018-11-29T11:32:37.263+06:30
#> 54    2018-11-29T10:08:12.488+06    2018-11-29T12:01:17.736+06
#> 57    2018-12-02T10:59:15.035+06    2018-12-02T11:32:11.290+06