Surveys that show start date earlier than first day of data collection

This function check that all interviews in the dataset start after the actual first day of data collection. There is an option to automatically mark for deletion the surveys which have started before the first day of data collection.

isSurveyStartedBeforeTheAssessment(
  ds = NULL,
  dates = NULL,
  survey_consent = NULL,
  start_collection = NULL,
  reportingcol = NULL,
  delete = NULL
)

Arguments

ds
dataset as a data.frame object

dates
fields as a list of string where the survey start and end date is stored (c('start_date','end_date'))

survey_consent
name as a string of the field in the dataset where the survey consent is stored

start_collection
date as a string of the first day of data collection ('yyyy-mm-dd')

reportingcol
columns as a list of string name from the dataset you want in the result (c('col1','col2',...))

delete
delete action to be done as a boolean (TRUE/FALSE)

Value

dst same dataset as the inputed one but with survey marked for deletion if errors are found and delete=TRUE

ret_log list of the errors found

Examples

{
ds <- HighFrequencyChecks::sample_dataset
dates <- c("survey_start","end_survey")
survey_consent <- "survey_consent"
start_collection <- "2018-11-11"
reportingcol <- c("enumerator_id","X_uuid")
delete <- FALSE

list[dst,ret_log,var,graph] <- isSurveyStartedBeforeTheAssessment(ds,
                                                                  dates,
                                                                  survey_consent,
                                                                  start_collection,
                                                                  reportingcol,
                                                                  delete)
head(ret_log, 10)
}

#> [1] enumerator_id X_uuid        survey_start 
#> <0 rows> (or 0-length row.names)