household dataset to be used as an example

This dataset comes from the Host Community Multi Sector Needs Assessment conducted in 2018 by IMPACT/REACH on behalf of the ISCG and funded by ECHO. This is a sample of the data collected in both Ukhia and Teknaf. For privacy purpose, the GPS coordinates are faked but are still located in the designated areas.

sample_dataset

Format

A data frame with 498 rows and 587 variables:

X.5

integer COLUMN_DESCRIPTION

instance_name

character COLUMN_DESCRIPTION

survey_date

character COLUMN_DESCRIPTION

survey_start

character COLUMN_DESCRIPTION

enumerator_org

character COLUMN_DESCRIPTION

enumerator_id

integer COLUMN_DESCRIPTION

upazilla_name

character COLUMN_DESCRIPTION

union_name

character COLUMN_DESCRIPTION

informed_consent

logical COLUMN_DESCRIPTION

survey_consent

character COLUMN_DESCRIPTION

consent_received.respondent_info.respondent_age

integer COLUMN_DESCRIPTION

consent_received.respondent_info.respondent_gender

character COLUMN_DESCRIPTION

consent_received.respondent_info.hh_head

character COLUMN_DESCRIPTION

consent_received.respondent_info.hh_gender

character COLUMN_DESCRIPTION

consent_received.respondent_info.hh_size

integer COLUMN_DESCRIPTION

consent_received.individual_information.repeat_introduction

logical COLUMN_DESCRIPTION

consent_received.individual_information.inividual_info_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.adult_female_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.boy_6_11_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.girl_6_11_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.boy_12_18_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.girl_12_18_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.individual_edu_aid_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.hh_head_number_count

integer COLUMN_DESCRIPTION

consent_received.individual_information.hhead_warning

logical COLUMN_DESCRIPTION

consent_received.individual_information.individual_group_end

logical COLUMN_DESCRIPTION

consent_received.wash.primary_watsource

character COLUMN_DESCRIPTION

consent_received.wash.primary_watsource_other

logical COLUMN_DESCRIPTION

consent_received.wash.improved_water_access

character COLUMN_DESCRIPTION

consent_received.wash.improved_water_access_other

logical COLUMN_DESCRIPTION

consent_received.wash.enough_water

character COLUMN_DESCRIPTION

consent_received.wash.hh_water_problem

character COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem

character COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.too_far

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.too_steep

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.available_sometime

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.dangerous

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.tastes_bad

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.water_badsmell

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.water_notclean

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.dont_know

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem.other

logical COLUMN_DESCRIPTION

consent_received.wash.water_collection_problem_other

character COLUMN_DESCRIPTION

consent_received.wash.water_source_time

character COLUMN_DESCRIPTION

consent_received.wash.water_collection_time

character COLUMN_DESCRIPTION

consent_received.wash.water_treatment

character COLUMN_DESCRIPTION

consent_received.wash.water_treatment_time

character COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method

character COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.aquatabs

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.pur_sachets

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.boiling

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.cloth_filters

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.household_filters

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.solar_disinfection

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method.other

logical COLUMN_DESCRIPTION

consent_received.wash.water_treatment_method_other

logical COLUMN_DESCRIPTION

consent_received.wash.soap_handwash

character COLUMN_DESCRIPTION

consent_received.wash.hh_defecation

character COLUMN_DESCRIPTION

consent_received.wash.hh_defecation_other

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access

character COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems

character COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.latrine_too_far

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.too_many_using

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.latrine_notclean

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.insufficient_water

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.bad_smell

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.open_defecation_near_lat

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.not_private

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.no_separation

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.route_not_safe

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.latrine_not_safe

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.no_lighting

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.dont_know

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems.other

logical COLUMN_DESCRIPTION

consent_received.wash.latrine_access_problems_other

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times

character COLUMN_DESCRIPTION

consent_received.wash.handwash_times.before_eating

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.before_cooking

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.after_defecation

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.before_breastfeeding

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.before_feeding_chid

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.after_child_bottom

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.before_prayer

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.hands_look_dirty

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.hands_feel_dirty

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.other

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.dntknow_prefer

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times.after_eating

logical COLUMN_DESCRIPTION

consent_received.wash.handwash_times_other

logical COLUMN_DESCRIPTION

consent_received.wash.visible_trash

character COLUMN_DESCRIPTION

consent_received.wash.visible_faeces

character COLUMN_DESCRIPTION

consent_received.wash.visible_waste

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.tin

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.tarpaulin

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.hay

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.brick_cemen

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.bamboo

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.wood

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material.other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_roof_material_other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.bricks

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.clay

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.bamboo

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.cement

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.tarpaulin

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.wood

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.tin

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material.other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_wall_material_other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material.bricks

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material.cement

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material.wood

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material.dirt

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material.other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_floor_material_other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.solar_light

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.electricity_grid

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_electricity_access

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_cyclone_training

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_flood_training

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.solar_lamp

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.portable_torch

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.kitchen_set

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.floor_mat

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.cooking_stove

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.blanket

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.mosquito_net

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.children_clothes

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.male_clothes

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.female_clothes

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.radio

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.umbrella

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.fuel

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.latrine_bathroom

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items.tubewell

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.non_food_items_other

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.cooking_fuel

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.cooking_fuel_other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.house_land_ownership

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.house_land_ownership_other

logical COLUMN_DESCRIPTION

consent_received.shelter_nfi.land_house_deed

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.rent_hosted

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.hh_agreement

character COLUMN_DESCRIPTION

consent_received.shelter_nfi.eviction

character COLUMN_DESCRIPTION

consent_received.ppi.hh_rooms

character COLUMN_DESCRIPTION

consent_received.ppi.hh_television

character COLUMN_DESCRIPTION

consent_received.ppi.hh_fans

character COLUMN_DESCRIPTION

consent_received.ppi.hh_phones

character COLUMN_DESCRIPTION

consent_received.ppi.hh_vehicle

character COLUMN_DESCRIPTION

consent_received.ppi.hh_agri_land

character COLUMN_DESCRIPTION

consent_received.food_security.food_crisis

logical COLUMN_DESCRIPTION

consent_received.food_security.less_preferred_expensive

integer COLUMN_DESCRIPTION

consent_received.food_security.borrowed_food

integer COLUMN_DESCRIPTION

consent_received.food_security.limit_portion_size

integer COLUMN_DESCRIPTION

consent_received.food_security.restrict_consumption

integer COLUMN_DESCRIPTION

consent_received.food_security.reduce_meal_numbers

integer COLUMN_DESCRIPTION

consent_received.food_security.eat_elsewhere

integer COLUMN_DESCRIPTION

consent_received.food_security.women_eat_less

integer COLUMN_DESCRIPTION

consent_received.food_security.men_eat_less

integer COLUMN_DESCRIPTION

consent_received.food_security.nofood_wholeday

integer COLUMN_DESCRIPTION

consent_received.food_security.food_list

logical COLUMN_DESCRIPTION

consent_received.food_security.cereals_tubers

integer COLUMN_DESCRIPTION

consent_received.food_security.pulses_nuts_seeds

integer COLUMN_DESCRIPTION

consent_received.food_security.vegetables

integer COLUMN_DESCRIPTION

consent_received.food_security.fruits

integer COLUMN_DESCRIPTION

consent_received.food_security.dairy

integer COLUMN_DESCRIPTION

consent_received.food_security.meat_fish

integer COLUMN_DESCRIPTION

consent_received.food_security.oil_fats

integer COLUMN_DESCRIPTION

consent_received.food_security.sweets

integer COLUMN_DESCRIPTION

consent_received.food_security.spices_condiments

integer COLUMN_DESCRIPTION

consent_received.food_security.market_access

character COLUMN_DESCRIPTION

consent_received.food_security.market_travel_time

character COLUMN_DESCRIPTION

consent_received.food_security.main_source_food

character COLUMN_DESCRIPTION

consent_received.food_security.main_source_food_other

logical COLUMN_DESCRIPTION

consent_received.food_security.food_hh_sustain

integer COLUMN_DESCRIPTION

consent_received.food_security.main_income

character COLUMN_DESCRIPTION

consent_received.food_security.main_income.non_agricultural

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.agricultural

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.domestic

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.petty_trade

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.small_business

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.large_business

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.skilled_labour

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.fishing

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.remittances

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.handicrafts

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.agricultural_production_sale

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.livestock_rearing

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.savings

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.begging

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.food

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.other_cash_assistance

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.sale_assistance

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.assistance_relative_friends

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.gather_sell_firewood

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.zakat

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.other

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income.none

logical COLUMN_DESCRIPTION

consent_received.food_security.main_income_other

logical COLUMN_DESCRIPTION

consent_received.food_security.hh_econ_status

character COLUMN_DESCRIPTION

consent_received.food_security.hh_living_standard

character COLUMN_DESCRIPTION

consent_received.food_security.hh_spending

logical COLUMN_DESCRIPTION

consent_received.food_security.spend_food

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_medication

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_education

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_fix_shelter

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_clothing

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_hygiene

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_fuel

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_hh_items

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_transport

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_communication

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_tobacco

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_rent

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_debts

integer COLUMN_DESCRIPTION

consent_received.food_security.spend_other

integer COLUMN_DESCRIPTION

consent_received.food_security.work_outside_country

character COLUMN_DESCRIPTION

consent_received.food_security.study_outside_country

character COLUMN_DESCRIPTION

consent_received.food_security.marry_outside_country

character COLUMN_DESCRIPTION

consent_received.food_security.return_outside_country

character COLUMN_DESCRIPTION

consent_received.food_security.forced_work_outside_country

character COLUMN_DESCRIPTION

consent_received.food_security.never_heard_back

character COLUMN_DESCRIPTION

consent_received.health.health_access

character COLUMN_DESCRIPTION

consent_received.health.health_access_barriers

character COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.health_facility_too_far

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.infrequent_services

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.treatment_unavailable

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.overcrowded_services

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.insufficient_poorquality_staff

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.not_adequate_material

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.dont_know_to_go

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.safety_concerns

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.too_expensive

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.dont_trust_doctors

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.face_discrimination

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers.other

logical COLUMN_DESCRIPTION

consent_received.health.health_access_barriers_other

logical COLUMN_DESCRIPTION

consent_received.health.new_health_facility

character COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility

character COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility.dont_know_about_services

logical COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility.services_too_far

logical COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility.servicesare_not_for_hc

logical COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility.prefer_services_already_exist

logical COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility.other

logical COLUMN_DESCRIPTION

consent_received.health.no_access_new_facility_other

logical COLUMN_DESCRIPTION

consent_received.health.comm_health_worker

character COLUMN_DESCRIPTION

consent_received.health.lactating_women

integer COLUMN_DESCRIPTION

consent_received.health.pregnant_women

integer COLUMN_DESCRIPTION

consent_received.health.clinic_visit_pregnancy

integer COLUMN_DESCRIPTION

consent_received.health.hh_marriage_person

character COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem

character COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.doctors

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.midwife_nurse

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.older_relatives

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.traditional_healer

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.other_mothers

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.friends

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.others

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.wouldnt_seek_help

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.not_applicable

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem.dont_know

logical COLUMN_DESCRIPTION

consent_received.health.women_breastfeed_problem_other

logical COLUMN_DESCRIPTION

consent_received.education.boy_prim_edu_barrier

character COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier

character COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.facillities_too_far

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.safety_concerns_way_facilities

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.services_too_expensive

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.services_too_crowded

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.insufficient_poorquality_staff

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.school_not_adequate_infra

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.school_not_adequate_material

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.face_discrimination

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.children_support_family

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.parents_think_not_appropriate

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.other

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier.dont_know

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_other

logical COLUMN_DESCRIPTION

consent_received.education.girl_prim_edu_barrier

character COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier

character COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.facillities_too_far

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.safety_concerns_way_facilities

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.services_too_expensive

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.services_too_crowded

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.insufficient_poorquality_staff

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.school_not_adequate_infra

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.school_not_adequate_material

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.face_discrimination

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.children_support_family

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.parents_think_not_appropriate

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.other

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier.dont_know

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_other

logical COLUMN_DESCRIPTION

consent_received.education.boy_second_edu_barrier

character COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1

character COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.facillities_too_far

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.safety_concerns_way_facilities

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.services_too_expensive

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.services_too_crowded

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.insufficient_poorquality_staff

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.school_not_adequate_infra

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.school_not_adequate_material

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.face_discrimination

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.children_support_family

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.parents_think_not_appropriate

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.other

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1.dont_know

logical COLUMN_DESCRIPTION

consent_received.education.boy_edu_barrier_1_other

logical COLUMN_DESCRIPTION

consent_received.education.girl_second_edu_barrier

character COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1

character COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.facillities_too_far

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.safety_concerns_way_facilities

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.services_too_expensive

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.services_too_crowded

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.insufficient_poorquality_staff

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.school_not_adequate_infra

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.school_not_adequate_material

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.face_discrimination

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.children_support_family

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.parents_think_not_appropriate

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.other

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1.dont_know

logical COLUMN_DESCRIPTION

consent_received.education.girl_edu_barrier_1_other

logical COLUMN_DESCRIPTION

consent_received.education.child_rights

character COLUMN_DESCRIPTION

consent_received.education.child_education

character COLUMN_DESCRIPTION

consent_received.education.edu_aid_material

character COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind

character COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind.health_hygiene_kit

logical COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind.winterization_kit

logical COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind.school_supply

logical COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind.other

logical COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind.none

logical COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind.dont_know

logical COLUMN_DESCRIPTION

consent_received.education.edu_aid_kind_other

character COLUMN_DESCRIPTION

consent_received.child_protection.missing_child

character COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk

character COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.violence_home

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.violence_community

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.child_labour

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.child_marriage

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.recruitment_armed_group

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.kidnapping

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.detention

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.sexual_abuse

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.psychological_trauma

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.no_registration_newborn

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.natural_hazards

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.dntknow_no_answer

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk.other

logical COLUMN_DESCRIPTION

consent_received.child_protection.boy_risk_other

character COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk

character COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.violence_home

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.violence_community

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.child_labour

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.child_marriage

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.recruitment_armed_group

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.kidnapping

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.detention

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.sexual_abuse

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.psychological_trauma

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.no_registration_newborn

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.natural_hazards

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.dntknow_no_answer

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk.other

logical COLUMN_DESCRIPTION

consent_received.child_protection.girl_risk_other

character COLUMN_DESCRIPTION

consent_received.child_protection.child_distress

character COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.withdrawn_family_friends

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.aggressive_outbursts

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.changes_appetite_eating_habits

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.headaches

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.recurrent_bedwetting

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.sleep_disturbances

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.upset_stomach_pain

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.new_recurring_fear

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.clinging

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.excessive_crying

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.behave_younger

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.startle_easily

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.substance_abuse

logical COLUMN_DESCRIPTION

consent_received.child_protection.child_distress.none

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_interact

character COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya

character COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.social_interactions

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.buying_goods_services

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.selling_goods_services

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.working_for_rohingya

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.hiring_rohingya_work

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.renting_to_rohingya

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.casual_interactions

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.other

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.none

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya.prefer_not_answer

logical COLUMN_DESCRIPTION

consent_received.child_protection.past_interaction_rohingya_other

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_comm_relation

character COLUMN_DESCRIPTION

consent_received.child_protection.opinion_rohingya

character COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel

character COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.competition_services

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.cultural_difference

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.unfair_distribution

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.competition_resources

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.competition_jobs

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.threat_crime

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.other

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.dont_know

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel.prefer_not_answer

logical COLUMN_DESCRIPTION

consent_received.child_protection.rohingya_feel_other

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups

character COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.health

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.education

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.safety_security

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.respond_natural_disaster

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.support_disabilities

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.protect_children

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.none

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups.other

logical COLUMN_DESCRIPTION

consent_received.child_protection.community_groups_other

character COLUMN_DESCRIPTION

consent_received.child_protection.hh_safe

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menst_material_access

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge.not_available_market

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge.preferred_type_not_available

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge.too_expensive

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge.other_needs_prioritized

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge.other

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge.dntknow_prefer

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.menstrual_challenge_other

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.gvb_questions_female

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_family_decision

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.contraception_decision

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_movement

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_market

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_health_center

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_neighborhood

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_religious_space

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.control_money

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.buy_vegetables_fruits

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.buy_clothes

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.buy_medicine

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.buy_toiletries

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_need_help

logical COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_shelter

character COLUMN_DESCRIPTION

consent_received.gender_section.women_question_consent.female_financial_support

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.gbv_questions_male

logical COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.family_decision_male

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.hh_chores_male

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.married_women_work

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.right_to_express

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.wife_beaten

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.son_daughter_school

character COLUMN_DESCRIPTION

consent_received.gender_section_1.male_question_consent.community_decision

character COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know

character COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.finding_miss_people

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.security_situation

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.regiter_aid

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_water

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_food

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_shelter

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.nutrition_information

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.food_prices

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.crop_livestock_prices

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_fuel

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.weather_natural_hazards

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_health_care

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_help

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.prevent_attack

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.replace_personal_documentation

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_access_education

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.find_work

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_transport

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.get_financial_support

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.about_rohingya

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.info_aid_agencies

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.aid_complain

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.bad_behaviour_aidworker

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.expected_aidworker_behaviour

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.other

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.none

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know.dntknow_no_answer

logical COLUMN_DESCRIPTION

consent_received.cwc.thing_to_know_other

character COLUMN_DESCRIPTION

consent_received.cwc.information_sources

character COLUMN_DESCRIPTION

consent_received.cwc.information_sources.tv

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.face_to_face

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.newspaper

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.radio

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.internet

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.facebook

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.whatsapp

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.sms

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.call

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.loudspeaker

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.community_leaders

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.religious_leaders

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.govt_official

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.police_army

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.aid_worker

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.community_meetings

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.other

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources.dntknow_prefer

logical COLUMN_DESCRIPTION

consent_received.cwc.information_sources_other

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information

character COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.tv

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.face_to_face

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.newspaper

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.radio

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.internet

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.facebook

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.whatsapp

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.sms

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.call

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.loudspeaker

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.community_leaders

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.religious_leaders

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.govt_official

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.police_army

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.aid_worker

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.community_meetings

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.other

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information.dntknow_prefer

logical COLUMN_DESCRIPTION

consent_received.cwc.accurate_information_other

character COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback

character COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.speak_service_provider

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.speak_community_leader

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.comm_meetings

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.helpline

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.complaint_form

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.suggestion_box

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.social_media

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.send_sms

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.send_email

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.send_letter

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.other

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback.dont_know

logical COLUMN_DESCRIPTION

consent_received.cwc.provide_feedback_other

character COLUMN_DESCRIPTION

consent_received.cwc.warning_message

character COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning

character COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.cyclone_preparedness

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.mosque_loudspeaker

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.signal_flag

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.radio

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.television

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.newspaper

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.internet

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.word_of_mouth

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.other

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.dont_know

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning.loudspeaker

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_early_warning_other

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future

character COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.cyclone_preparedness

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.mosque_loudspeaker

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.signal_flag

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.radio

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.television

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.newspaper

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.internet

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.word_of_mouth

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.other

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.dont_know

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future.loudspeaker

logical COLUMN_DESCRIPTION

consent_received.cwc.receive_message_future_other

logical COLUMN_DESCRIPTION

end_survey

character COLUMN_DESCRIPTION

end_note

logical COLUMN_DESCRIPTION

gps_reading

character COLUMN_DESCRIPTION

X_gps_reading_latitude

double COLUMN_DESCRIPTION

X_gps_reading_longitude

double COLUMN_DESCRIPTION

X_gps_reading_altitude

double COLUMN_DESCRIPTION

X_gps_reading_precision

integer COLUMN_DESCRIPTION

meta.instanceID

character COLUMN_DESCRIPTION

X_id

integer COLUMN_DESCRIPTION

X_uuid

character COLUMN_DESCRIPTION

X_submission_time

character COLUMN_DESCRIPTION

X_index

integer COLUMN_DESCRIPTION

X_parent_table_name

logical COLUMN_DESCRIPTION

X_parent_index

character COLUMN_DESCRIPTION

X_tags

integer COLUMN_DESCRIPTION

X_notes

character COLUMN_DESCRIPTION

X

character COLUMN_DESCRIPTION

X.1

double COLUMN_DESCRIPTION

X.2

double COLUMN_DESCRIPTION

X.3

double COLUMN_DESCRIPTION

X.4

integer COLUMN_DESCRIPTION

Source

http://somewhere.important.com/